Rename namespace from "Ngrok" to "Tunnel"
