+++ src/src/Environment/Environments/E2E/E2EEnvironment.php
@@ -7,8 +7,8 @@ use QIT_CLI\Environment\Docker;
 use QIT_CLI\Environment\Environments\Environment;
 use QIT_CLI\Environment\EnvUpChecker;
 use QIT_CLI\Environment\PluginActivationReportRenderer;
-use QIT_CLI\Ngrok\NgrokConfig;
-use QIT_CLI\Ngrok\NgrokRunner;
+use QIT_CLI\Tunnel\NgrokConfig;
+use QIT_CLI\Tunnel\NgrokRunner;
 use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
